add 593

+++ LeetCode/629_K个逆序对数组.cpp
@@ -17,8 +17,8 @@ public:
     int kInversePairs(int n, int k) {
         vector<vector<int>> f(2, vector<int>(static_cast<size_t>(k) + 1));
         f[0][0] = 1;
-        for (size_t i = 1; i <= n; ++i) {
-            for (size_t j = 0; j <= k; ++j) {
+        for (size_t i = 1; i <= static_cast<size_t>(n); ++i) {
+            for (size_t j = 0; j <= static_cast<size_t>(k); ++j) {
                 size_t cur = i & 1, prev = cur ^ 1;
                 f[cur][j] = (j - 1 >= 0 ? f[cur][j - 1] : 0) - (j - i >= 0 ? f[prev][j - i] : 0) + f[prev][j];
                 if (f[cur][j] >= mod) {
